+++ src/Network/Network/ChangeLog.md
@@ -19,6 +19,16 @@
 --->
 
 ## Upcoming Release
+* Added new cmdlets to replace old product name `virtual router` with new name `route server` in the future.
+    - `New-AzRouteServer`
+    - `Get-AzRouteServer`
+    - `Remove-AzRouteServer`
+    - `Update-AzRouteServer`
+    - `Get-AzRouteServerPeer`
+    - `Add-AzRouteServerPeer`
+    - `Update-AzRouteServerPeer`
+    - `Remove-AzRouteServerPeer`
+    - Added deprecation attribute warning to the old cmdlets.
 * Updated format list and format table views for Get-AzVirtualNetworkGatewayConnectionIkeSa
 
 ## Version 4.5.0